Bets now lock in. Payouts are shown.
When a user places a bet it locks the user in by disabling their bet
buttons. If the user page loads and a bet is present in their table it
will also disable the buttons.
Also payouts are calculated on the page as odds and betting changes. It
is also shown what their payout will be after they bet and lock in their
odds.